requirements.txt
# Core Dependencies - Latest compatible versions
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.1.0
langchain-chroma>=0.1.0

# Agent Framework
langgraph>=0.0.20

# Vector Store & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.0.0

# OpenAI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
gradio>=4.0.0

# Text Processing
tiktoken>=0.5.0

# Optional: For advanced features
rank-bm25>=0.2.0
scikit-learn>=1.3.0

# Web Application Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# MongoDB Database
pymongo>=4.6.0
motor>=3.3.0
